change test to .ts instead of .js
