+++ src/App.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import { IonApp } from '@ionic/react';
 
 test('renders without crashing', () => {
-  const { baseElement } = render(<App />);
+  const { baseElement } = render(<IonApp />);
   expect(baseElement).toBeDefined();
 });
