+++ src/App.tsx
@@ -23,7 +23,6 @@ import '@ionic/react/css/display.css';
 /* Theme variables */
 import './theme/variables.css';
 
-
 const App: React.FC = () => {
   return (
     <IonApp>
@@ -33,7 +32,7 @@ const App: React.FC = () => {
         </IonRouterOutlet>
       </IonReactRouter>
     </IonApp>
-  )
+  );
 };
 
 export default App;
