+++ src/components/actions.ts
@@ -1,12 +1,6 @@
+import { ActionTypes } from './enums';
 import { store } from './../index';
 import { ILocation } from './location';
-
-export enum ActionTypes {
-  SHOW_LOCATION = 'SHOW_LOCATION',
-  SHOW_ALL = 'SHOW_ALL',
-  HIDE_LOCATION = 'HIDE_LOCATION',
-  SHOW_FILTERED = 'SHOW_FILTERED'
-}
 
 export interface IBaseAction {
   type: ActionTypes;
