+++ src/global.d.ts
@@ -1,4 +1,5 @@
-export {}
+/* eslint-disable */
+export {};
 declare global {
     interface Window {
         __REDUX_DEVTOOLS_EXTENSION__?: any;
