+++ src/index.tsx
@@ -8,14 +8,14 @@ import { Provider } from 'react-redux';
 
 export const store = createStore(
   locationApp,
-    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
 
 ReactDOM.render(
-    <Provider store={store}>
+  <Provider store={store}>
     <App />
   </Provider>,
-document.getElementById('root')
+  document.getElementById('root')
 );
 
 // If you want your app to work offline and load faster, you can change
